requirements.txt
streamlit
streamlit-option-menu
python-dotenv
pandas
crewai
crewai_tools
langchain-community
langchain-groq